Move the `env-server` module under the `@bayou` namespace.
