+++ local-modules/@bayou/env-server/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "env-server",
+  "name": "@bayou/env-server",
   "version": "1.0.0",
 
   "dependencies": {
