+++ local-modules/app-setup/package.json
@@ -13,7 +13,7 @@
     "@bayou/deps-util-common": "local",
     "@bayou/doc-common": "local",
     "doc-server": "local",
-    "env-server": "local",
+    "@bayou/env-server": "local",
     "@bayou/hooks-server": "local",
     "@bayou/ot-common": "local",
     "@bayou/see-all": "local",
