+++ local-modules/doc-server/package.json
@@ -8,7 +8,7 @@
     "@bayou/app-common": "local",
     "@bayou/codec": "local",
     "@bayou/doc-common": "local",
-    "env-server": "local",
+    "@bayou/env-server": "local",
     "file-store": "local",
     "@bayou/file-store-ot": "local",
     "@bayou/hooks-server": "local",
