+++ local-modules/file-store-local/LocalFileStore.js
@@ -6,7 +6,7 @@ import afs from 'async-file';
 import path from 'path';
 
 import { Codec } from '@bayou/codec';
-import { Dirs } from 'env-server';
+import { Dirs } from '@bayou/env-server';
 import { BaseFileStore, FileCache } from 'file-store';
 import { TheModule as fileStoreOt_TheModule } from '@bayou/file-store-ot';
 import { Logger } from '@bayou/see-all';
