+++ local-modules/testing-server/package.json
@@ -6,7 +6,7 @@
     "@bayou/deps-ansi-color": "local",
     "@bayou/deps-testing-common": "local",
     "@bayou/deps-testing-server": "local",
-    "env-server": "local",
+    "@bayou/env-server": "local",
     "@bayou/testing-common": "local",
     "@bayou/util-common": "local"
   }
