+++ server/package.json
@@ -20,7 +20,7 @@
     "client-bundle": "local",
     "@bayou/deps-util-server": "local",
     "dev-mode": "local",
-    "env-server": "local",
+    "@bayou/env-server": "local",
     "@bayou/hooks-server": "local",
     "@bayou/promise-util": "local",
     "@bayou/see-all": "local",
